refactor(core): Reorganize parameter manager source tree and sync docs
