+++ README.md
@@ -26,7 +26,7 @@ It is designed for projects that need a clean way to:
 
 ### Quick start
 
-1. Add the core sources from `src/` to your build.
+1. Add `src/par.c` and the sources from `src/def`, `src/layout`, `src/persist`, and `src/port` to your build. Add `src/persist/backend` only when you use the packaged backend adapter.
 2. Provide a project-specific `par_table.def` at the package root.
 3. Provide `port/par_cfg_port.h` in your include path.
 4. Optionally provide `port/par_if_port.c` and `port/par_atomic_port.h` when your platform needs them.
@@ -69,31 +69,37 @@ static void app_init(void)
 parameters/
 ├── README.md
 ├── CHANGE_LOG.md
-├── doc/
-│   └── DeviceParameter_VerificationReport.xlsx
 ├── docs/
+│   ├── DeviceParameter_VerificationReport.xlsx
 │   ├── api-reference.md
 │   ├── architecture.md
 │   └── getting-started.md
 ├── src/
-│   ├── backend/
-│   │   ├── par_store_backend.h
-│   │   └── par_store_backend_gel_nvm.c
 │   ├── par.c
 │   ├── par.h
-│   ├── par_atomic.h
-│   ├── par_bitwise_impl.inc
 │   ├── par_cfg.h
-│   ├── par_def.c
-│   ├── par_def.h
-│   ├── par_if.c
-│   ├── par_if.h
-│   ├── par_layout.c
-│   ├── par_layout.h
-│   ├── par_nvm.c
-│   ├── par_nvm.h
-│   ├── par_storage_init.inc
-│   └── par_typed_impl.inc
+│   ├── def/
+│   │   ├── par_def.c
+│   │   ├── par_def.h
+│   │   ├── par_id_map_static.c
+│   │   └── par_id_map_static.h
+│   ├── detail/
+│   │   ├── par_bitwise_impl.inc
+│   │   ├── par_storage_init.inc
+│   │   └── par_typed_impl.inc
+│   ├── layout/
+│   │   ├── par_layout.c
+│   │   └── par_layout.h
+│   ├── persist/
+│   │   ├── backend/
+│   │   │   ├── par_store_backend.h
+│   │   │   └── par_store_backend_gel_nvm.c
+│   │   ├── par_nvm.c
+│   │   └── par_nvm.h
+│   └── port/
+│       ├── par_atomic.h
+│       ├── par_if.c
+│       └── par_if.h
 └── template/
     ├── par_cfg_port.htmp
     ├── par_layout_static.htmp
@@ -111,7 +117,7 @@ This repository contains the reusable module core and templates. A real integrat
 
 ### Optional, depending on configuration
 
-- `port/par_if_port.c` when `PAR_CFG_IF_PORT_EN = 1` and the target needs stronger platform hooks than the weak defaults in `par_if.c`
+- `port/par_if_port.c` when `PAR_CFG_IF_PORT_EN = 1` and the target needs stronger platform hooks than the weak defaults in `src/port/par_if.c`
 - `port/par_atomic_port.h` when `PAR_ATOMIC_BACKEND = PAR_ATOMIC_BACKEND_PORT`
 - generated static layout header when `PAR_CFG_LAYOUT_SOURCE = PAR_CFG_LAYOUT_SCRIPT`
 - a concrete storage backend implementation when `PAR_CFG_NVM_EN = 1`
@@ -124,13 +130,14 @@ This repository contains the reusable module core and templates. A real integrat
 
 ## Key integration notes
 
-- `par_cfg.h` includes `par_cfg_port.h` unconditionally, so your build must provide that header.
+- `src/par.h` is the main public entry header. Keep `parameters/src` on the compiler include path so application code can use `#include "par.h"`.
+- `par_cfg.h` includes `par_cfg_port.h` unconditionally, so your build must provide that header and make its directory visible to the compiler.
 - `PAR_CFG_ENABLE_TYPE_F32` controls whether floating-point parameter support and the related typed APIs are compiled in.
 - `PAR_CFG_ENABLE_RUNTIME_VALIDATION` and `PAR_CFG_ENABLE_CHANGE_CALLBACK` control whether normal setters include runtime validation callbacks and on-change callbacks.
 - The module separates **internal parameter enumeration** (`par_num_t`) from **external parameter IDs** (`id`).
 - The current ID lookup implementation uses a one-entry-per-bucket hash map generated at compile time from `par_table.def`. External IDs must therefore be not only unique, but also collision-free under the configured hash geometry. Optional runtime diagnostic scans can be enabled with `PAR_CFG_ENABLE_RUNTIME_ID_DUP_CHECK` and `PAR_CFG_ENABLE_RUNTIME_ID_HASH_COLLISION_CHECK` when additional startup logs are useful. See `docs/architecture.md` for the collision rule and avoidance guidance.
 - Unchecked setter APIs skip runtime validation callbacks and on-change callbacks, so they should be reserved for tightly controlled hot paths. Bitwise fast setters are further restricted to `U8` / `U16` / `U32` flags or bitmask parameters. Legacy `*_fast()` names remain as deprecated aliases.
-- NVM support is optional. When enabled, `par_nvm.c` depends on a mounted storage backend interface and on ID and persistence metadata being enabled. The package can build the `GeneralEmbeddedCLibraries/nvm` adapter from `src/backend/`, or the application can provide its own `par_store_backend_get_api()` implementation.
+- NVM support is optional. When enabled, `src/persist/par_nvm.c` depends on a mounted storage backend interface and on ID and persistence metadata being enabled. The package can build the `GeneralEmbeddedCLibraries/nvm` adapter from `src/persist/backend/`, or the application can provide its own `par_store_backend_get_api()` implementation.
 - `par_init()` applies startup default values directly to live storage. Integer default values from `par_table.def` are compiled into a grouped width-based storage object, while `F32` default values are applied to the 32-bit storage group after layout offsets are available only when `PAR_CFG_ENABLE_TYPE_F32 = 1`. Because this startup initialization does not go through the public setter path, it does not invoke runtime validation or on-change callbacks.
 - `PAR_CFG_ENABLE_RESET_ALL_RAW` controls whether raw reset-all support and grouped default mirror snapshot support are enabled.
