+++ src/detail/par_bitwise_impl.inc
