+++ src/detail/par_storage_init.inc
