+++ template/par_layout_static.htmp
@@ -1,6 +1,6 @@
 /**
  * @file par_layout_static.h
- * @brief 
+ * @brief Declare the generated static layout contract.
  * @author wdfk-prog ()
  * @version 1.0
  * @date 2026-03-21
@@ -17,7 +17,7 @@
 
 #include <stdint.h>
 
-#include "../src/par_def.h"
+#include "def/par_def.h"
 
 #ifdef __cplusplus
 extern "C" {
